Refactor: split render methods into a source file
